add and use LinkTree component
