+++ src/app/page.tsx
@@ -1,39 +1,6 @@
-import { AiFillGithub, AiFillLinkedin } from 'react-icons/ai';
-import { FiCodepen } from 'react-icons/fi';
-
-import { URLS } from '@/data/constants';
-
-import List from '@/components/List';
-import WebLink from '@/components/WebLink';
+import LinkTree from '@/components/LinkTree';
 
 import styles from './page.module.scss';
-
-const items = [
-  {
-    id: 1,
-    child: (
-      <WebLink href={URLS.GITHUB}>
-        <AiFillGithub />
-      </WebLink>
-    ),
-  },
-  {
-    id: 2,
-    child: (
-      <WebLink href={URLS.CODEPEN}>
-        <FiCodepen />
-      </WebLink>
-    ),
-  },
-  {
-    id: 3,
-    child: (
-      <WebLink href={URLS.LINKEDIN}>
-        <AiFillLinkedin />
-      </WebLink>
-    ),
-  },
-];
 
 export default function Home() {
   return (
@@ -43,8 +10,7 @@ export default function Home() {
           <h1>Hey! I&apos;m Charlotte and I love building websites.</h1>
           <p>Making things move on the computer screen still sparks joy.</p>
           <hr />
-          <h2>Find Me</h2>
-          <List items={items} />
+          <LinkTree />
         </div>
       </section>
     </main>
