+++ src/components/LinkTree/index.tsx
@@ -1,0 +1,3 @@
+import LinkTree from './LinkTree';
+
+export default LinkTree;
